functions/long-poller/requirements.txt
crowdstrike-foundry-function==1.1.1
crowdstrike-falconpy
PyJWT
